added the ticktock feature
This will add the hidden joke timer to the extension. This will activate by clicking on the tick-tock button. After clicking it opens a new tab and has the presidential timer. The presidential timer shows how long we have left in our current presidency. This addition was what my mother suggested and Mr.Walker says take peoples suggestions so that is why we are adding this piece.